Move utils.ts to field-types/utils/render-from-elements.ts
