+++ packages/dataviews/src/field-types/boolean.tsx
@@ -12,7 +12,7 @@ import type {
 	FieldTypeDefinition,
 	NormalizedField,
 } from '../types';
-import { renderFromElements } from '../utils';
+import renderFromElements from './utils/render-from-elements';
 import { OPERATOR_IS, OPERATOR_IS_NOT } from '../constants';
 
 function sort( a: any, b: any, direction: SortDirection ) {
